Renamed unit_conversion_adapter to unit_conversion
