+++ chatterbot/logic/__init__.py
@@ -6,7 +6,7 @@ from .multi_adapter import MultiLogicAdapter
 from .no_knowledge_adapter import NoKnowledgeAdapter
 from .specific_response import SpecificResponseAdapter
 from .time_adapter import TimeLogicAdapter
-from .unit_conversion_adapter import UnitConversion
+from .unit_conversion import UnitConversion
 
 __all__ = (
     'LogicAdapter',
